Trying to add project details into api
